Bug#36262779: MySQL fails to execute query if a subquery refer to a column of parent table
The problem is with resolving a column in an outer query block.
See test case in test.subquery_bugs for reference.
The column to be resolved is "ref_0.a" from the innermost query block.
It should resolve to table "ref_0" (aliased from "t1") in the outermost
query block. But the query block that is immediately outer to the
innermost one selects the same column "ref_0.a", and the column in
question is resolved against this. Since this is in the select list,
it is also checked that it is not farther out in the select list than
the subquery that contains the column reference that we are resolving.
This check fails, and hence we give the error ER_ILLEGAL_REFERENCE:
Reference 'c1' not supported (forward reference in item list).

The reason for this is that we check in the select list for the
column "ref_0.a". However, the SQL standard only supports checking
the select list for column references that are simple identifiers.
In this case, the column reference contains both a table name and
a column table, and should thus only be checked against columns in
appropriate tables.

The solution is therefore to restrict checking in select list to
simple column names. The change is performed once in
Item_field::fix_outer_field() and once in Item_ref::fix_fields().
In addition, it was necessary to replace an Item_ref constructor
in Item_field::fix_outer_field() with a different one that does not
require a following fix_fields() call, because this particular
fix_fields() could rely upon a lookup using table name and column name
in the select list, which has just been made invalid by the above
change.

It was also possible to remove an obsolete Item_ref constructor.

The change has some impact to queries in the test suite, but nothing
serious. Some tests had a change in error message from
ER_BAD_FIELD_ERROR to ER_WRONG_FIELD_WITH_GROUP because a field in
a HAVING clause is now actually found in a table in an outer scope.
Other tests had a change in error from ER_WRONG_FIELD_WITH_GROUP
to ER_BAD_FIELD_ERROR because a field in HAVING clause used to be
erroneously found in the select list.

Change-Id: I69f0c14a24ebce0fd5f868f27b544a551d69659d