Bug#36446730 GSL log messages do not specify the owner
Problem:

The GSL is acquired by the binlog, the metadata change monitor and the
user session threads. When the GSL is acquired or released, a message
is logged. It is not obvious who has acted on it, much less if the THD
is of a background thread,
unless a correlation effort is made.

Analysis:

If the MySQL server system is using performance schema
interface (PSI), all ndbcluster plugin threads and user session
threads created are registered on the performance schema. This allows
the usage of the PFS thread id to correlate instrumentation data from
the multiple PFS tables. This PFS thread id is not observed anywhere,
only the THD thread id in the logged message.

Solution:

In order to identify who has acquired it, this commit adds the thread
name to the Ndb_Thd and, on each GSL obtained, a info string is
produced fetching the PFS thread id.

The log level of GSL messages was changed to "info" so that they are
always produced. A user using defaults, now has no need to specify
ndb-extra-logging and still benefit from feedback.

The log messages from GSL subsystem now looks like this:
  <time> 2 [System] [MY-010865] [NDB] Global schema lock acquired
(name=ndb_binlog, pfs_thread_id=30)
  <time> 2 [System] [MY-010865] [NDB] Global schema lock release

Or, from a user session thread:
  <time> 14 [System] [MY-010865] [NDB] Global schema lock acquired
(pfs_thread_id=297)
  <time> 14 [System] [MY-010865] [NDB] Global schema lock release

Change-Id: I07a6a816ba1f6ea0d9c734193e4af7946e342d11