Bug#36487681 - mysql server 8.3.0 crashed at mysqlx_generate_document_id
Description
===========
Passing "null" values as parameters to "MYSQLX_GENERATE_DOCUMENT_ID"
udf function, crashes server. To reproduce the issue the "null" value
must returned by a function, for example:

  JSON_CONTAINS_PATH(@doc, 'one', '$.b') # where @doc is null
  IF(@expr, 1, NULL)

Combining:
  SELECT MYSQLX_GENERATE_DOCUMENT_ID(IF(1=2,1,null));

Fix
===
The values can't be verified at "udf_init" calls, because the may be
evaluated just before calling the "udf_callback". Because of this
limitation, we can't report a wrong usage to the user, instead
we must use default values when the value is set to NULL.

Change-Id: I4ab4b5321b33b2ed8f4ea352b1e3130a5b755707