WL#16268[3]: Support for InnoDB Partitions Phase 2 - Enrich DD Info
After a mysqld crash, we need to automatically reload only the
partitions that were previously loaded. For this purpose, we enrich the
Data Dictionary information with a per-partition secondary_load flag.
This flag is exposed in the partitions table of Information Schema.

Moreover, we ensure that we log an entry in binlog for all partition
load/unload operations but we don't update the table's secondary_laod
flag in DD in case of a partition unload that does not unload all the
partitions

Other changes:
- add partition load/unload test cases when the mock secondary engine
  is used

Change-Id: If1196d983eb8e6f7954bedb3871064675574b1e3