Bug#36388981 AT: testNodeRestart -n ChangeNumLDMsNR T1 D1 fails in .2ndbd due to signal
Problem:
Test ChangeNumLDMsNR changes the number of LDM threads by setting
AutomaticThreadConfig and increasing the numCPUs (that internally
increases the number of LDM to use). That works fine for multi
threads data nodes (ndbmd) but for  the single thread case (ndbd)
data nodes crashes in DIH with division by zero.
During the READ_CONFIG_INFO handling in DIH, if classic fragmentation
is off, there is a clause to check if the number of LDMs configured
via ATC+numCPUs is enough to ensure that the resulting  number of
partitions per LDM is not larger than 16. In the calculus of partitions
per LDM the use of the variable globalData.ndbMtLqhWorkers, initialized
to 0 causes the division by zero.

if ((c_fragments_per_node_ * cnoReplicas) / globalData.ndbMtLqhWorkers >
MAX_FRAG_PER_LQH) {
  ....

This happens because for single threaded data node the ndbMtLqhWorkers
variable is not set during the initial configuration setup
( Configuration::setupConfiguration() ).

Solution:
- Ignore AutomaticThreadConfig and ClassicFragmentarion configuration
since, by design, they are 'exclusive' for ndbmtd.
This fixes the test because failing scenario is only in non-classic
fragmentation path.

- Skip test for ndbd scenario, it is useless in that case.

Change-Id: Ic0496226fa4d3e2d5f68c6bbdbbb38ffcb663330