Bug#35658072 management nodes are unable to allocate node id to data and sql pods when restarted
Current MGMD code will attempt to match a connecting client's
address to a configured address of the given type.

This involves resolving the configured address to an IP address
and checking whether it matches the socket level address presented
by the client.
This is intended to :
 - Give some control over which hosts different configured nodes
   can connect from depending on network / naming control.
 - Give some optional automatic identity determination where any
   node of type T connecting from host H is assumed to be [one of]
   node[s] N1[, N2, ..Nx].
 - Allow some node identities to connect from anywhere, while others
   must connect from one location.
 - ?

Configuration can include :
 - No address        : Anyone can connect as that nodeid
 - An IP address     : Only connections from that IP can connect
                       as that nodeid
 - A logical address : Only connections from hosts
                       resolving-at-connect-time to that IP can
                       connect as that nodeId.

Indirections can occur at the IP layer, but this mechanism only really
understands indirection via a logical address where a name resolution
service is involved in obtaining IP addresses for logical addresses.

The mechanism originally assumed that all logical addresses in
configuration would remain resolvable for the entire lifecycle
of a cluster.

When running in dynamic environments, name to IP mappings can be
missing or stale.  Underlying IPs may be reused with different
mappings over time.
Cluster nodes may stop and be restarted with a different IP not
yet resolved in the name->IP mapping.

A previous fix extended the mechanism to recognise when some name->IP
mappings were missing, and treat this as an indicator of a dynamic
environment, where client address match failure should be treated
as a temporary condition, rather than a permanent misconfiguration.
This causes standard client code in NdbApi and the data nodes to
retry a number of times before giving up, which may give sufficient
time for the name->address mapping to resolve and avoid a visible
process initialisation error.

This fix goes further and treats all client address match failures
as temporary conditions, so that even in cases where all configured
names are resolvable, and none of them match, the client will retry
a number of times.

Since the special case of unresolvable hosts is no longer special,
it is removed.

This change means that actual misconfiguration or incorrect usage
will not fail immediately, but will fail after a number of retries.

Change-Id: Iec20a70f0634a35f457f40b91e1193eeeb1bce7d