Bug#36533080: Remove the field cached_table from class Item_ident
Removed the field cached_table from class Item_ident.
Replaced its use with a new field m_table_ref in the same class.
Also removed the field table_ref from class Item_field.
The field Item_ident::cached_table was used to reference a base table if
the column reference was against a base table, and it was used to
reference a view or derived table if the column reference was against
a view or derived table. The same purpose is now handled by m_table_ref.

Some obsolete code in Item_field::fix_fields() that was used for
repeated resolving when cached_table was set, was removed. This also
made it possible to remove the "outer_fixed" local variable.

Commented-out code in transform_table_subquery_to_join_with_derived()
was also deleted.

The field m_cached_table in class copy_field_info was also obsolete
and could be deleted.

Change-Id: I51565ad744060d3342e8fbb881fd1e32f652db08