Bug#35189564: hypergraph_optimizer switch is not set to its Default value
Setting optimizer_switch to DEFAULT didn't change
hypergraph_optimizer=on back to hypergraph_optimizer=off, even though
the default is off.

This unusual behaviour was introduced in order to ensure test coverage
for the hypergraph optimizer with "mtr --hypergraph" even in test
cases that contained "SET optimizer_switch = DEFAULT".

This patch makes "SET optimizer_switch = DEFAULT" behave as one would
expect, and reset all optimizer switches, including
hypergraph_optimizer, to the default value.

To preserve test coverage, "mysqltest --hypergraph" is changed to
enable the hypergraph optimizer both in the session and in the global
scope. This way, if a test resets @@session.optimizer_switch to
DEFAULT, it takes the default from the @@global.optimizer_switch and
preserves hypergraph_optimizer=on.

A consequence of this change is that new connections opened by
"mysqltest --hypergraph" will also use the hypergraph optimizer.
Before, only the default connection used the hypergraph optimizer.
Many tests now use the hypergraph optimizer more than before.
Especially those that open multiple connections and those that perform
a restart of the server. A lot of tests are therefore updated to
accept differences between the old optimizer and the hypergraph
optimizer.

Test changes are limited to those test suites that have already been
prepared to run cleanly with "mtr --hypergraph" in mysql-trunk. This
means the tests that are in the suites that are run by default by mtr,
in Debug and RelWithDebInfo configurations with default cmake options.
The changes also cover the tests in those suites that run only with
the --ps-protocol or --big-test options.

Change-Id: I132ed126d52d3ad8cf8cdf0a1718ef3922ee1ce0