BUG#36548269 Log when incidents occur to mysqld.log
Log before and after when ndbcluster plugin writes an incident to
binlog.

There are several functions that write incident messages to the
injector, doing more or less the same but with some variation. In order
to be able to add logging, verification and resource management a new
utility function is extracted for writing an incident message to the
injector. Since function is used rarely and only in error scenarios make
sure to write a message to the log both before and after writing to
injector. Furthermore change so that write of incident is not
infinitely retried when failure to inject incident to binlog occurs.

This will end up in the MySQL Server error log like this:
  <time> 2 [System] [MY-010866] [NDB] Binlog: Writing incident 'cluster
    disconnect' to binlog
 <time> 2 [ERROR] [MY-010853] [Server] cluster disconnect An incident
   event has been written to the binary log which will stop
   the replicas.
 <time> 2 [System] [MY-010866] [NDB] Binlog: Incident 'cluster
   disconnect' written to binlog

Change-Id: Iaaed28fe6c2254eca3802421922632e8c1807b8a