Bug#36549567 Allow ndb_sign_keys user to specify CA lifetime
The utility ndb_sign_keys is able to create a CA, but does not
allow the user to specify the valid lifetime of the CA. All CA
certificates are created with a validity of four years.

This patch adds a new option --CA-days to ndb_sign_keys. The
default value of the option corresponds to four years, matching
the existing behavior.

Change-Id: I63dec1bd91fea5c74a86a1e7b6b27aa1c71ab507