Bug#30845983 NDB.NDB_DD_RESTORE_COMPAT FAILS OFTEN IN PB2 WEEKLY
Problem:
--------

The test fails due to expected table upgrade errors due to the
presence of columns with obsolete data types. The errors are
suppressed in one MySQL Server but also appear in the second
MySQL Server especially on slower machines.

Fix:
----

Suppress the errors in the second MySQL Server as well.

Change-Id: Iab8dc05aa26b10a2375216f5f8cfdfeba465b25f