Bug#36457756: Invalid result converting string to numeric
When a character string is converted to a numeric value, any non-numeric
data trailing the numeric value should cause an error with strict mode
and a warning with non-strict mode. However, due to a bad pointer
calculation in Item_field::val_xxx() functions, some invalid strings
will not cause errors or warnings. Notice that this is a legacy error,
however the length and character set of the string may impact what
causes an error and what does not.

The pointer arithmetic is cleaned up for three val_xxx() functions:
val_int(), val_real() and val_decimal(). Of these, only val_int() and
val_real() were prone to errors, but val_decimal() had the similar
treatment for consistency.

Also cleaned up a few more pointer cast issues in same component.

Change-Id: I7420981c4a5a9ba2e4f0a2bd436d4b90a2d68ade