Bug#36556725: Failure in Group_check::do_ident_check Bug#36557029: Recent failure in check_column_grant_in_table_ref
In the fix for bug#36533080, a code line had to be added in
find_order_in_list(), due to an assignment to the field
Item::m_table_ref in find_field_in_tables(). But this caused a unit
test to fail, and more compensation code had to be added. This in turn
caused the two errors reported here.

However, instead of adding more compensation code, it was researched
that the assignment in find_field_in_tables() was redundant, since
the same would happen in Item_field::set_field(). Even the assignment
to field_index is redundant and could be removed.

The fix is thus to remove the redundant assignments, as well as the
compensation code.

Another redundant setting of m_table_ref in a constructor for class
Item_field was also removed.

Change-Id: I79f2c7dd0ab5deec240a96bac987c39b558e0f7f