WL#16142: Enable use of cursors from stored procedures with sec engines
This fix enables use of cursors from stored procedures with Rapid
secondary storage engine.

The functionality is enabled with statements like these from a
stored procedure:

  DECLARE mycursor CURSOR FOR SELECT c1 FROM t;
  OPEN mycursor;
  FETCH mycursor INTO c1_value;
  ...
  CLOSE mycursor;

It is not possible to use cursors with prepared statements
against the Rapid storage engine from connectors like Connector/C.

Result sets for cursors are stored in temporary tables, like when
cursors are evaluated in the regular MySQL execution engine, and
the same rules and limitations apply.

The test rapid.sp is adjusted to show that cursor-based statements are
offloaded to the Rapid storage engine.

Bug#36157680: Cursor query with incorrect fetch variable hangs

The problem is an infinite loop in execution of cursor FETCH operations
in stored procedures, when execution in the RAPID execution engine is
enabled.

The problem occurs during execution against the primary execution
engine. The FETCH operation reports an error because the number of
fetch variables is inconsistent with number of select expressions in
the SELECT statement associated with the cursor. This makes program
execution enter the function check_secondary_engine_statement()
for the iparent CALL statement. Note that CALL can only execute in
the primary engine, however the execution of FETCH returns with
secondary execution indicated
(secondary_engine_optimization = SECONDARY). This makes the server retry
execution repeatedly until it fails.

The solution to this problem is to ensure that procedure execution
returns with secondary_engine_optimization = PRIMARY_ONLY, since we
only allow execution of procedures in the primary engine anyway.

Further, we ensure in the function open_secondary_engine_tables() that
when offload is not possible (as it is for e.g a CALL statement),
engine optimization is set to PRIMARY_ONLY. The consequence of this is
that the server will never attempt to retry a procedure execution that
ended in an error.

Bug#36186967: Failure in temptable::Block::destroy

This bug occurs because a temporary table is not cleaned up properly
after statement execution. The reproduction case for this problem has
a DISTINCT aggregator, which is attached to an Item_sum object that is
created in the statement's mem_root. After first optimization against
the primary engine, another optimization against the secondary engine
is attempted. Between optimizations, it is assumed that all remains of
the first optimization is cleaned out. However, this does not happen,
and the temporary table associated with the aggregator is never deleted.

Bug#36225802: Cursor query crashed when table is unloaded and
              procedure called

This bug is also related to missing cleanup after optimization.
But here the problem is missing cleanup of the item list: Because it
is not cleaned up between optimizations, references to mem_root objects
that were wiped out when the execution mem_root was recycled are kept
and will reference deleted memory. Thus, when the final cleanup of the
routine statement is performed, we traverse the item list and encounter
deallocated memory.

The solution to these two problems is to clean up both the execution
mem_root and the statement mem_root, and the associated item lists,
after an optimization that will be retried, as well as the LEX
associated with the statement that also contains a list of plugins used
by the statement.

In addition, a small refactoring of the main execution loop in
sp_lex_instr::validate_lex_and_execute_core() is performed. The loop
is executed as long as the criteria for retry are fulfilled. The loop
can be exited either with successful or error completion. In case of
a retry, a common section at the end of the loop clears the diagnostics
area, cleans all relevant mem_roots and item lists, and disables the
general log.

A regression test case for bug#36225802 has been added, but it proved
difficult to create a small enough similar test case for Bug#36186967.

Change-Id: I49af00cd240ac005934ef3a0b4dd912e3264b2c8