Bug#36563430 router does a last_check_in update in an INVALIDATED Cluster, when restarted
The Router uses the INVALIDATED cluster status to stop the
RW client traffic. It does not use that information internally
to prevent the periodic last_check_in updates. By default for
the ClusterSet the update frequency is '0' so no updates are done
except the initial one - hence the issue is most likely to be
visible when the Router gets restarted with the invalidated
primary Cluster.

This patch fixes the issue, now the Router will not do the
last_check_in update on the invalidated Cluster.

Change-Id: I0829053d308a86ff2754428c658bf84e2f22da3d