Bug#36568964 Test concurrent_grants fails
The destructor for NDB_SCHEMA_OBJECT makes several assertions about
the state of the schema object. However, the state is protected by
a mutex, and the destructor does not acquire this mutex before testing
the state. This could explain the rare intermittent failure seen in
this bug report.

Acquire the mutex in the destructor.

Change-Id: I4b8568f491244854b2295535308d680cd931a98c