Bug#36521175 Cluster/J tests fail with Connector/Java 8.0.23
In testClusterj, improve the classpath checking, and use the fixed
value of GMT-3 for the time zone in the JVM.

The fix for bug#28818191 "FIX TIME ZONE ISSUES IN NDB.CLUSTERJ TEST
IN WINDOWS" opted to start the JVM using the -Duser.timezone=$TZ
option, coercing the JVM to use a fixed time zone whenever the
cluster/j test suite is run from MTR. In practice, for me, this
means that the server runs in GMT, and the JVM runs in GMT-3.
This patch applies this same fix to testClusterj.

I will add a note about the Cluster/J test suite. In the Connector/J
developer guide, sec. 6.6.1 "Preserving Time Instants", it is stated
that *only* the TIMESTAMP data type is designed to preserve instants.
The TIME, DATE, and DATETIME types do not have this ability. Yet, it
seems that a set of the Cluster/J "read NDB write JDBC" tests are
based on the premise that TIME and DATETIME are able to preserve
instants.  This could include TimeAsUtilDate, TimeAsSqlTimestamp, and
similarly DatetimeAsSqlTimestamp and DatetimeAsUtilDate, plus the
"Query" versions of these tests. The fix here and in 28818191 could be
seen as a work-around that masks this issue. On the other hand, to the
extent that the real purpose of the tests is to verify that the
various data types can be encoded and decoded without corruption,
such workaround and masking might be appropriate.

If a user reports that a MySQL query and a Cluster/J app report
differing values for a TIME or DATETIME column, they should be
referred to this section of the Connector/J guide.

Change-Id: If1b819e51694eb554c8396cf81c6df2324bc3ce2