Bug#36562972: Character set changing unexpectedly
When resolving a call to the function REPLACE, the character set and
collation of the function result are copied from the first argument.
The remaining two arguments, if they are literal values, are converted
to this character set. However, due to a mistake, only the second
argument was converted. The third argument was not.

This fix ensures the third argument is converted too.

Change-Id: Iafebde65b5c58a53f6b9c131f883de9d9e597a0b