Bug#36170903 The current_memory for performance_schema.table_handles suddenly went up
Problem
=======

In production, the memory used by the performance schema can suddenly
spike up to the maximum theoretical size of a PFS_buffer_scalable_container.

This was observed in particular for the underlying scalable buffer
used for table performance_schema.table_handles,
which can consume 1024 entries or 1024 rows each,
for a total memory of 9 Gb.

When such spike occurs, the content of the buffer appears empty.

Root cause
==========

The problem was never reproduced,
the root cause is found by code analysis instead.

The issue is a race condition in the implementation of
class PFS_buffer_scalable_container,
when the buffer is expanded to add a new page.

The code takes great care to use atomic operations
to detect that a new page was added, for example in:

```
      /* (2-a) Atomic Load, array= m_pages[current_page_count] */
      array = m_pages[current_page_count].load();
```

The problem is that ensuring the array pointer itself is seen with
atomic operations is not sufficient.

Members of the array, for example:
- PFS_buffer_default_array::m_full
- PFS_buffer_default_array::m_max
are not atomic.

As a result, it is possible that:

Thread 1:
- creates a new page,
- sets m_full=false
- sets m_max=1024
- adds the new page in m_pages[current_page_count]

Thread 2:
- find the new page in m_pages[current_page_count]
- reads m_full and finds true
- reads m_max and finds 0

When such race occurs, thread 2 considers the newly added page as full.

Repeating the process leads to creation of all pages,
and to the memory consumption reported.

Fix
===

For several threads to cooperate correctly on:
- class PFS_buffer_default_array
- class PFS_buffer_scalable_container
every member accessed concurrently is changed to use atomic operations.

Change-Id: I7e4d25428c2353ad7757e893e6400e73ff1709e3