Bug#36418426: Even though there is an index, the index is not being used enough when LEFT JOIN.
Issue:
Even though there is an index under the left join conditions,
only the columns of the equi-join condition used the index
and the columns in the range condition are filtered.

Analysis:
Index condition pushdown was not getting implemented
even though when all the fields in the join condition are
part of an index because the whole join condition was
wrapped around a trigger condition 'IS_NOT_NULL_COMPL'.
Because of this filtering was done for the range conditions.

Fix:
Check if the above mentioned trigger condition could be used
for ICP in case of outer joins.

Change-Id: I8e962ad530a82da0aa1c57a590a09696bd674a64