Bug#36379879: DETERMINISTIC FUNCTIONS can return incorrect results
This issue might happen when we have a function containing a JOIN with
ON clause inside the return statement. If that SELECT needs a reprepare
due to a table metadata change (Eg: Due to flush tables in this case) between
two executions, we may see the ON clause getting lost.

The reason for this is that, in order for the SP instruction to be
reprepared, the text string is first prefixed with 'SELECT '. Thus,
the parser will parse this as a SELECT statement. However, the resolver
will see a query block structure that is unexpected, and not call
the function Query_block::apply_local_transforms() which (among other
things) handles the ON clause and merges into the WHERE clause, hence
the missing ON clause. The resolver expects to see the SQL command
SQLCOM_END (rather than SQLCOM_SELECT).

This can be achieved by implementing adjust_sql_command() function
for class sp_lex_instr (which implements the RETURN statement), and
adjust the SQL command type to SQLCOM_END.

Change-Id: I3d7b98a20d7813db5e13f4475074de5f82c7dd85