Bug#36528707 miss notification when write larger than 1MB size data
The logic in notify_about_advanced_write_lsn() was not prepared to
handle writes long enough to cause a wrap-around and the same event
slot being assigned to first and last lsn in the written range - it
would wrongly interpret it as whole range falling into the same slot
and thus miss to notify threads waiting for events from all the
other slots along the way. This wouldn't be a critical problem as
eventually log notifier would notify them anyway, but might cause
unnecessary delay, which would be difficult to track down as the
logic looks superficially correct.

The fix employed is to check that the range is short enough that
it couldn't cause a wrap-around.

We thank Zongzhi Chen (Alibaba Taobao) for this contribution!

Change-Id: Iba40a6cfb117aa9c9ff98f2868f812dfb634f793