Bug#36590692 main.subquery_scalar_to_derived crashes with --ps-protocol
During the prepare phase of a prepared statement, Item::transform
would call Query_expression::replace_items. This method, upon
replacement of an item with another, unconditionally calls
THD::change_item_tree, which should now only be used during the
optimization phase, not during the prepare phase.

This lead to two kinds of symptoms: in one case, the prepare mem_root
of change records inserted into THD::change_list were deallocated and
thrashed before rollback_item_tree_changes was executed, so we would
access garbage.  In another case, we would hit an assert that
THD::change_list be empty after a prepare, which it was not in this
case. Both cases are added to the tests.

During normal execution (not prepared statement), the call does
nothing anyway, so the error wasn't seen without prepared statements
(or --ps-protocol as reported in this issue).
Query_expression::replace_items belongs to the transform processor, so
no need to call change_item_tree; Item::transform should only be used
during the prepare phase for permanent transforms.

Solution: replace the call to change_item_tree with a plain
assignment.

Change-Id: Ib52589b736b979407e8b81206b9153da89b89325