Bug#35352161: Mysql dies when a functional index on FIND_IN_SET is added
During CREATE/ALTER, we use Create_field instead of Field.
Create_field_wrapper has-a Create_field, and is-a Field.
This way, Create_field_wrapper can stand in for a Field
for methods that expect one.

We try to cast this Create_field_wrapper (which is-a Field)
to Field_enum (which is-a Field_string, which is-a Field).
The reason for this is that we wish to look up the typelib
from the Field_enum.

This typecast understandably fails, as while both classes
are subclasses of Field, they don't exist on a direct path
of inheritance. This bug is also facilitated by
Create_field_wrapper passing our check

  if (field->real_type() == MYSQL_TYPE_SET) {

since the Create_field_wrapper::type() returns the contained
Create_field's ->sql_type. (I.e. we try to make sure we've
got a Field_set, and the wrapper pretends to be one -- but
it isn't one, and does not have the typeset data, which is
what we're ultimately looking for.)

To improve the situation, we now check in the affected
method whether the field is in truth a Create_field_wrapper
and if so, don't try to acquire the typelib.

Change-Id: I0d5c3b9af7956eb8322ac35194f3af73195c468f