Bug#35189564: hypergraph_optimizer switch is not set to its Default value
Post-push fix.

main.bug39022 fails when run with "mtr --hypergraph --no-skip",
because the test is marked as skipped because a statement has been
annotated with --skip_if_hypergraph. The --no-skip mechanism doesn't
understand that it's expected that it's skipped with hypergraph.

Fixed by skipping the entire test, not just the single statement, when
running with --hypergraph. That case is handled fine by an exception
in mysql-test/include/excludenoskip.list.

Change-Id: Ic425b6c228f55318c4512f66dbf91580c1d35d16