Bug#36532705 - Limit Regular Statement Handle usage.
The Regular Statement Handle facilitates a session to execute a
query from the server within an isolated setup. It also retains
the result-set of a query, if present, in-memory. Allowing
numerous concurrent uses of Regular Statement Handles could
exhaust resources. Therefore, this patch introduces a limitation
on the number of Regular Statement Handles.

Regular Statement Handles are primarily used by the server, and
users are generally unaware of their usage. To limit Regular
Statement Handle usage, a hard-coded session limit of 1024 is
introduced as part of this patch.

Change-Id: I09e7d2f4fb43dff7d6f81487289215656557a012