Bug#35710213: mysql server 8.1.0 failed at Item_field::used_tables
The original fix for this problem caused a regression in test
i_main.fulltext-bug17865492 in --ps-protocol mode.
The reason for this is a missing implementation of the function
update_used_tables() for the class Item_func_match. When rebuilding
used tables information and item properties, the argument "against" is
never considered. This means that for execution of prepared statements,
the Item_func_match expression is no longer considered to be an
aggregated expression, and the optimizer will pick the wrong
substitution field when setting up an intermediate temporary table.

The fix for this problem is to implement update_used_tables() for this
class, and explicitly add used tables information and properties for
the "against" argument.

Change-Id: Ic3a4c61ae6706a772a2c3616927a08751a3ade42