Bug#36079456 Assertion `having_cond->has_subquery() || !(having_cond->used_tables() & ~(1 | PSEUDO_TABLE_BITS))' failed.
Server exit not seen in production builds, although even those are not
functioning correctly for this case.  In transforming subqueries to
derived tables (WL#12885), replacement of a subquery in a HAVING
condition failed to use an Item_ref, instead using an Item_field. This
leads to the assert seen cf. this quote in
JOIN::add_having_as_tmp_table_cond:

  "Pushing parts of HAVING to an internal temporary table.
   Fields in HAVING condition may have been replaced with fields in an
   internal temporary table. This table has map=1. "

For the above to work, the item replacing the subquery must be an item
ref, so the slice system can do its thing: allowing the having
condition to reference the temporary table, not the original derived
table (which has a map entry of 4 in the simplified repro, not 1 as
expected, hence the assert).

The code did not correctly detect that the subquery was part of a
HAVING condition. This is regression after commit

  ca2cd90bffb WL#9384: Prepare each DML statement once

although due to another error it didn't manifest for the repro until
later, in commit

  ca5275ff6e7 Bug#35060385 Item::update_used_tables should also
              update the PROP_SUBQUERY - for transforms

Change-Id: Iaddce7fddd7e08ebbb6fb0e683f81fc1a00f0d89