Bug#35352161: Mysql dies when a functional index on FIND_IN_SET is added
Post-push fix.

The test case for the bug fails with the hypergraph optimizer because
the plan shown by an EXPLAIN statement is different. The expected plan
is an index range scan, but the hypergraph optimizer chooses a table
scan.

Fixed by making the test table bigger, so that a table scan looks less
attractive and both optimizers prefer an index range scan. There is
still a difference between the plan chosen by the two optimizers; the
old optimizer puts a redundant filter on top of the index range scan,
whereas the hypergraph optimizer doesn't. This is fixed by using
EXPLAIN FORMAT=JSON and extracting only the interesting part of the
plan.

Change-Id: Ief90f6857c75c98f0e44c722c31a34a22c66bc3b