Bug#36624279 component tests fail to shutdown properly
Component tests that use:

  lauch_router(..., -1s)

may run into a hang on process shutdown if the signal handler thread
isn't started before the process manager tries to terminate the Router
with SIGTERM.

If SIGTERM is sent before the signal-handler is ready, the signal is
ignored and the router keeps running, leading to the failed test.

This scenario can sometimes be triggered by running the router tests
in parallel on a overloaded machine.

Change
======

- use router_spawner().wait_for_sync_point(RUNNING) to wait for the
  signal handler.

Change-Id: I9a62ce3822f900305265ba496b872c5c3787d9e3