Bug#36624363 component_gr_state fails on slow machines
routertest_component_gr_state fails sometimes on slow machines due to
more than one metadata-refresh leading to a "node skipped" line in the log.

The test tries to ensure that the only one metadata-refresh triggers the
no-quorum state. But on slow machines a metadata-refresh right between
the update of the mock-servers on the new cluster-state may get the
cluster into a inconsistent state and trigger a 2nd "node skipped" line.

Background
==========

The test first updates the PRIMARY to the new state and then the
SECONDARYs that the router is supposed to switch to.

If the refresh happens after the PRIMARY has been updated and before
SECONDARYs got their update, then the cluster is inconsistent.

Change
======

- first update the SECONDARYs to the new state, then the PRIMARY as the
  SECONDARYs will only be checked AFTER the PRIMARY gets ruled out as
  "no-quorum"

Change-Id: I71cbbbbb523c11ec141423aa31e757d2f320525a