Bug#36624427 component_socket_close fails sporadically
routertest_component_socket_close fails spordically on slow machines
in

  SocketCloseTest/StaticRoundRobinUnixSocket

The test expects a test-internal "UnixSocketUser" to bind to a port and
that returns "error 1130" to the client, but instead the connection is
dropped which triggers "error 2013".

Background
==========

After a destination fails to connect, the router checks if a destination
is still not reachable to bring the destination back to the load-balancing.

The test forces a connection failure, binds the UnixSocketUser to that port
and expects the the UnixSocketUser's error-code. But on slow machines
the destination status check may also connect to the that UnixSocketUser
which only expects one connection, but not two.

Change
======

- rewrite the UnixSocketUser to expect multiple connections (the test
  and the destination-status).

Change-Id: I9a62ab8171143888f790017ddb087c2816f5c6be